change folder structure
